+++ tests/Webhook/Handler/SaleRefundedTest.php
@@ -9,12 +9,18 @@ namespace Swag\PayPal\Test\Webhook\Handler;
 
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStateHandler;
 use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionStates;
+use Swag\PayPal\RestApi\V1\Api\Payment\Transaction\RelatedResource\Sale;
+use Swag\PayPal\RestApi\V1\PaymentStatusV1;
+use Swag\PayPal\RestApi\V1\Resource\SaleResource;
+use Swag\PayPal\Test\Mock\PayPal\Client\_fixtures\V1\GetResourceSaleResponseFixture;
 use Swag\PayPal\Test\Mock\Repositories\OrderTransactionRepoMock;
 use Swag\PayPal\Webhook\Handler\SaleRefunded;
 use Swag\PayPal\Webhook\WebhookEventTypes;
 
 class SaleRefundedTest extends AbstractWebhookHandlerTestCase
 {
+    private Sale $sale;
+
     public function testGetEventType(): void
     {
         $this->assertEventType(WebhookEventTypes::PAYMENT_SALE_REFUNDED);
@@ -24,6 +30,13 @@ class SaleRefundedTest extends AbstractWebhookHandlerTestCase
     {
         $webhook = $this->createWebhookV1();
         $this->assertInvoke(OrderTransactionStates::STATE_REFUNDED, $webhook, OrderTransactionStates::STATE_PAID);
+    }
+
+    public function testInvokeWithPartialRefund(): void
+    {
+        $webhook = $this->createWebhookV1();
+        $this->sale->setState(PaymentStatusV1::PAYMENT_PARTIALLY_REFUNDED);
+        $this->assertInvoke(OrderTransactionStates::STATE_PARTIALLY_REFUNDED, $webhook, OrderTransactionStates::STATE_PAID);
     }
 
     public function testInvokeWithoutParentPayment(): void
@@ -46,9 +59,15 @@ class SaleRefundedTest extends AbstractWebhookHandlerTestCase
 
     protected function createWebhookHandler(): SaleRefunded
     {
+        $saleResource = $this->createMock(SaleResource::class);
+        $this->sale = (new Sale())->assign(GetResourceSaleResponseFixture::get());
+        $this->sale->setState(PaymentStatusV1::PAYMENT_REFUNDED);
+        $saleResource->method('get')->willReturn($this->sale);
+
         return new SaleRefunded(
             $this->orderTransactionRepository,
-            new OrderTransactionStateHandler($this->stateMachineRegistry)
+            new OrderTransactionStateHandler($this->stateMachineRegistry),
+            $saleResource,
         );
     }
 }